+++ pennylane_cirq/__init__.py
@@ -15,7 +15,8 @@
 Plugin overview
 ===============
 """
-from .simulator_device import SimulatorDevice, MixedStateSimulatorDevice, QSimDevice
+from .simulator_device import SimulatorDevice, MixedStateSimulatorDevice
+from .qsim_device import QSimDevice
 
 from .ops import BitFlip, PhaseFlip, PhaseDamp, AmplitudeDamp, Depolarize
 from ._version import __version__
